[UI] 켈퍼 구현-2
